backend/requirements.txt
#Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

#Database & Caching
sqlalchemy==2.0.23
redis==5.0.1
python-dotenv==1.0.0

#AI & NLP
google-generativeai==0.3.2
textblob==0.17.1
nltk==3.8.1
openai==1.3.7

#Audio
pydub==0.25.1
requests==2.31.0

#Analytics & Reporting (Optional but useful)
matplotlib==3.8.0
pandas==2.1.1
numpy<2.0

#Utilities
pyyaml==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1

#authentication
google-auth-oauthlib==1.2.0
python-jose[cryptography]==3.3.0
requests-oauthlib==1.3.1